requirements.txt
pillow>=10.0
yt-dlp>=2024.1
pymupdf>=1.23
